# docs/api/openapi.json
{
...
  "info": {
    "title": "ITDO ERP4 API",
    "version": "0.1.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/leave-requests": {
      "get": {
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      },
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "endDate": {
                    "format": "date",
                    "type": "string"
                  },
                  "endTime": {},
                  "hours": {
                    "minimum": 0,
                    "type": "number"
                  },
                  "leaveType": {
                    "type": "string"
                  },
                  "notes": {
                    "type": "string"
                  },
                  "startDate": {
                    "format": "date",
                    "type": "string"
                  },
                  "startTime": {},
                  "userId": {
                    "type": "string"
                  }
                },
                "required": [
                  "userId",
                  "leaveType",
                  "startDate",
                  "endDate"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/leave-requests/{id}/submit": {
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/leave-requests/leader": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "userId",
            "required": false,
            "schema": {
              "minLength": 1,
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "pending_manager"
                  ],
                  "type": "string"
                },
                {
                  "enum": [
                    "approved"
                  ],
                  "type": "string"
                },
                {
                  "enum": [
                    "rejected"
                  ],
                  "type": "string"
                }
              ]
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "maximum": 300,
              "minimum": 1,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
  }
}